requirements.txt
selenium
pytest
webdriver-manager
Pillow
